Add s3 storage to our system so we can actually host images
