+++ amplify/backend.ts
@@ -12,25 +12,8 @@ import { stripePayment } from "./functions/stripe-payment/resource";
 import { nashWebhook } from "./functions/nash-webhook/resource";
 import { seedDevelop } from "./functions/seed-develop/resource";
 import { secret } from '@aws-amplify/backend';
-import { Duration } from 'aws-cdk-lib';
-import { Function } from 'aws-cdk-lib/aws-lambda';
 import { PolicyStatement } from 'aws-cdk-lib/aws-iam';
-import { defineStorage } from '@aws-amplify/backend-storage';
-
-// Define S3 storage for restaurant images
-export const storage = defineStorage({
-  name: 'restaurantImages',
-  access: (allow) => ({
-    'restaurant/*': [
-      allow.guest.to(['read']),
-      allow.authenticated.to(['read', 'write', 'delete'])
-    ],
-    'menu/*': [
-      allow.guest.to(['read']),
-      allow.authenticated.to(['read', 'write', 'delete'])
-    ]
-  })
-});
+import { storage } from './storage/resource';
 
 export const backend = defineBackend({
   auth,
